requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astroid==2.12.13
astunparse==1.6.3
autopep8==2.0.1
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
Django==4.1.3
django-guest-user==0.5.2
djangorestframework==3.14.0
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
isort==5.11.4
joblib==1.2.0
keras==2.11.0
lazy-object-proxy==1.8.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.2
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pylint==2.15.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
